Once again, very in-progress. Adding the first skeleton of a votingapi_action_handler. Each handler will be expected to implement three operations:
process -- accept a bundle of vote-result data and return a true or false.

fields -- accept a $comparison_value and return a collection of formapi elements used to edit them.

save -- accept a collection of form elements and return a string (or serialized collection of data) to be saved as a $comparison_value